Code/Python/requirements.txt
appdirs==1.4.4
attrs==23.1.0
Automat==20.2.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
bs4==0.0.1
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
chromedriver==2.24.1
click==8.0.3
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cssselect==1.2.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.0
fake-useragent==1.4.0
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.6.4
incremental==21.3.0
isodate==0.7.2
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
MarkupSafe==2.0.1
more-itertools==8.10.0
msal==1.31.1
msal-extensions==1.2.0
netifaces==0.11.0
oauthlib==3.2.0
outcome==1.3.0.post0
parse==1.20.0
portalocker==2.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pyee==8.2.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
requests-html==0.10.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
SecretStorage==3.3.1
selenium==4.16.0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
systemd-python==234
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.18
validators==0.22.0
w3lib==2.1.2
wadllib==1.3.6
websockets==10.4
wget==3.2
wsproto==1.2.0
zipp==1.0.0
zope.interface==5.4.0
